# Improve parsing performance