--- a/url/searchparams.go
+++ b/url/searchparams.go
@@ -159,6 +159,11 @@ func (s *SearchParams) Iterate(f func(pair *NameValuePair)) {
 
 func (s *SearchParams) String() string {
 	output := strings.Builder{}
+	toGrow := 0
+	for _, nvp := range s.params {
+		toGrow += (len(nvp.Name) * 3) + (len(nvp.Value) * 3) + 2 // *3 worst case: every rune is percent-encoded; 2 for '=' and '&';
+	}
+	output.Grow(toGrow)
 	for idx, nvp := range s.params {
 		if idx > 0 {
 			output.WriteRune('&')