Hi, I'm trying to replace github.com/ada-url/goada with whatwg-url, and I noticed that whatwg-url is a lot slower than goada in many cases.

I made a bunch of performance optimizations based on flame graphs, and saw 30%~100% performance improvements in various cases.

---
Before:

```
BenchmarkParse/1-12       303860              3725 ns/op            1864 B/op         74 allocs/op
BenchmarkParse/2-12       255351              4731 ns/op            2448 B/op        113 allocs/op
BenchmarkParse/3-12       698196              1665 ns/op             864 B/op         41 allocs/op
BenchmarkParse/4-12       830721              1314 ns/op             744 B/op         31 allocs/op
BenchmarkParse/5-12       585558              1856 ns/op            1176 B/op         59 allocs/op
BenchmarkParse/6-12       688225              1590 ns/op            1080 B/op         50 allocs/op
BenchmarkParse/7-12       395816              2993 ns/op            1704 B/op         71 allocs/op
BenchmarkParse/8-12       422863              2827 ns/op            1480 B/op         62 allocs/op
BenchmarkParse/9-12       327033              3358 ns/op            1816 B/op         76 allocs/op
...
```

After:

```
BenchmarkParse/1-12       496786              2181 ns/op            1512 B/op         53 allocs/op
BenchmarkParse/2-12       355843              2958 ns/op            1912 B/op         86 allocs/op
BenchmarkParse/3-12      1313612               921.3 ns/op           600 B/op         26 allocs/op
BenchmarkParse/4-12      1598416               746.9 ns/op           544 B/op         19 allocs/op
BenchmarkParse/5-12      1178409              1007 ns/op             576 B/op         32 allocs/op
BenchmarkParse/6-12      1390153               873.8 ns/op           520 B/op         25 allocs/op
BenchmarkParse/7-12       564979              1921 ns/op            1384 B/op         53 allocs/op
BenchmarkParse/8-12       643351              1851 ns/op            1400 B/op         54 allocs/op
BenchmarkParse/9-12       634221              1950 ns/op            1400 B/op         54 allocs/op
...
```