@yzqzss 2025-07-10
> What's the reason for your change from goada?

I see Zeno is a major adopter of goada. Sorry to let you know that we are trying to change from goada. :see_no_evil: 
The main reason is that we want to avoid CGO to make cross-compilation easier.

For example, I have not figured out how to cross-compile goada for arm64 MacOS on amd64 Linux.

![iamge](https://github.com/user-attachments/assets/cb6d9a21-1cd4-4fb4-b579-555a590d5ff8)